pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitrepodb"
dynamic = ["version"]
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "click",
    "GitPython",
    "PyGithub",
    "python-dotenv",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.7",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
gitrepodb = "gitrepodb.gitrepodb:gitrepodb"

[tool.hatch.version]
path = "gitrepodb/__about__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/gitrepodb"
]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
